backend/requirements.txt
asgiref==3.7.2
astroid==3.0.1
autopep8==1.5.5
dill==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
flake8==3.8.4
importlib-metadata==6.8.0
isort==5.12.0
jedi==0.17.2
mccabe==0.6.1
packaging==23.2
parso==0.7.1
platformdirs==4.0.0
pluggy==1.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
pycodestyle==2.6.0
pydocstyle==6.3.0
pyflakes==2.2.0
pylint==3.0.2
python-dotenv==1.0.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytoolconfig==1.2.6
pytz==2023.3.post1
rope==1.11.0
snowballstemmer==2.2.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
ujson==5.8.0
yapf==0.40.2
zipp==3.17.0
